fix:  the order of I/O points with same timestamp.
Since the handling of an I/O point is quick,
the write point and read point of same message in two nodes may have
same timestamp.
And when collecting fault sequences from traces,
the read points may be put before the write point.
I add a function to scan fault sequence and fix this.