feat: add debugger package to Phosphor module
When developing, it is useful to print some debug information.
I add a package of debugger.
This package could contain all debuggers for different aims.
It use a separated DebuggerConf to control whether opening a debugger.
With project development, it is important to remove useless debuggers.
Debuggers should cooperate with loggers.